+++ tests/integration/test_sharepoint.py
@@ -1,6 +1,9 @@
 import pytest
 import os
+import pathlib
+import json
 import pandas as pd
+import configparser
 from viadot.exceptions import CredentialError
 
 from viadot.sources import Sharepoint
@@ -8,10 +11,13 @@ from viadot.config import local_config
 from viadot.task_utils import df_get_data_types_task
 from viadot.tasks.sharepoint import SharepointToDF
 
+from prefect.tasks.secrets import PrefectSecret
+
 
 def get_url():
-    credentials = local_config.get("SHAREPOINT")
-    return credentials["url"]
+    with open(".config/credentials.json", "r") as f:
+        config = json.load(f)
+    return config["SHAREPOINT"]["url"]
 
 
 @pytest.fixture(scope="session")
@@ -45,8 +51,9 @@ def test_connection(sharepoint):
 
 def test_sharepoint_to_df_task():
     task = SharepointToDF()
+    credentials_secret = PrefectSecret("SHAREPOINT_KV").run()
     res = task.run(
-        credentials_secret="sharepoint-aia-account",
+        credentials_secret=credentials_secret,
         sheet_number=0,
         path_to_file="Questionnaires.xlsx",
         url_to_file=get_url(),