+++ viadot/sources/cloud_for_customers.py
@@ -130,6 +130,7 @@ class CloudForCustomers(Source):
         Returns:
             List: List of dictionaries.
         """
+
         metadata_url = self.change_to_meta_url(url)
         column_maper_dict = self.map_columns(metadata_url)
         entity_list = []
@@ -147,6 +148,7 @@ class CloudForCustomers(Source):
         return entity_list
 
     def map_columns(self, url: str = None) -> Dict[str, str]:
+
         """Fetch metadata from url used to column name map.
 
         Args: