+++ viadot/tasks/sqlite.py
@@ -2,6 +2,7 @@ from logging import log
 from typing import Any, Dict
 
 import pandas as pd
+from pendulum import instance
 import prefect
 from prefect import Task
 from prefect.utilities.tasks import defaults_from_attrs