+++ tests/integration/test_sharepoint.py
@@ -10,6 +10,7 @@ from viadot.sources import Sharepoint
 from viadot.config import local_config
 from viadot.task_utils import df_get_data_types_task
 from viadot.tasks.sharepoint import SharepointToDF
+
 from prefect.tasks.secrets import PrefectSecret