+++ viadot/sources/cloud_for_customers.py
@@ -120,6 +120,7 @@ class CloudForCustomers(Source):
             return self._to_records_other(url=url)
 
     def response_to_entity_list(self, dirty_json: Dict[str, Any], url: str) -> List:
+
         """Changing request json response to list.
 
         Args: