+++ viadot/flows/sharepoint_to_adls.py
@@ -2,6 +2,7 @@ import os
 from typing import Any, Dict, List
 import pendulum
 from prefect import Flow, task
+from pathlib import Path
 from prefect.backend import set_key_value
 from prefect.utilities import logging
 
@@ -35,6 +36,7 @@ class SharepointToADLS(Flow):
         output_file_extension: str = ".csv",
         local_dir_path: str = None,
         adls_dir_path: str = None,
+        adls_file_name: str = None,
         adls_sp_credentials_secret: str = None,
         if_empty: str = "warn",
         if_exists: str = "replace",
@@ -55,6 +57,7 @@ class SharepointToADLS(Flow):
             output_file_extension (str, optional): Output file extension - to allow selection of .csv for data which is not easy to handle with parquet. Defaults to ".csv".
             local_dir_path (str, optional): File directory. Defaults to None.
             adls_dir_path (str, optional): Azure Data Lake destination folder/catalog path. Defaults to None.
+            adls_file_name (str, optional):Name of file in ADLS. Defaults to None.
             adls_sp_credentials_secret (str, optional): The name of the Azure Key Vault secret containing a dictionary with
             ACCOUNT_NAME and Service Principal credentials (TENANT_ID, CLIENT_ID, CLIENT_SECRET) for the Azure Data Lake.
             Defaults to None.
@@ -81,13 +84,19 @@ class SharepointToADLS(Flow):
         else:
             self.local_file_path = self.slugify(name) + self.output_file_extension
         self.local_json_path = self.slugify(name) + ".json"
-        self.adls_schema_file_dir_file = os.path.join(
-            adls_dir_path, "schema", self.now + ".json"
-        )
         self.adls_dir_path = adls_dir_path
-        self.adls_file_path = os.path.join(
-            adls_dir_path, self.now + self.output_file_extension
-        )
+        if adls_file_name != None:
+            self.adls_file_path = os.path.join(adls_dir_path, adls_file_name)
+            self.adls_schema_file_dir_file = os.path.join(
+                adls_dir_path, "schema", Path(adls_file_name).stem + ".json"
+            )
+        else:
+            self.adls_file_path = os.path.join(
+                adls_dir_path, self.now + self.output_file_extension
+            )
+            self.adls_schema_file_dir_file = os.path.join(
+                adls_dir_path, "schema", self.now + ".json"
+            )
 
         super().__init__(*args, name=name, **kwargs)