+++ CHANGELOG.md
@@ -18,16 +18,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - KeyVault support in Sharepoint task
 - added KeyVault support in `CloudForCustomers` tasks
 
-- 
 ### Changed
 - pinned Prefect version to 0.15.11
 - `df_to_csv` now creates dirs if they don't exist
+- `ADLSToAzureSQL` - when data in csv coulmns has unnecessary "\t" then removes them
 
 ### Fixed
 - fixed an issue with duckdb calls seeing initial db snapshot instead of the updated state (#282)
 - C4C connection with url and report_url optimization
 - column mapper in C4C source
-
 
 ## [0.2.15] - 2022-01-12
 ### Added