+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 ## [Unreleased]
+### Added
+- Added `adls_file_name` in  `SupermetricsToADLS` and `SharepointToADLS` flows
 ## [0.3.2] - 2022-02-17
 ### Fixed
 - fixed an issue with schema info within `CheckColumnOrder` class.